+++ constants.py
@@ -6,7 +6,7 @@ EXTENSION_NAME = "VRML2 Material Studio"
 EXTENSION_VERSION = (0, 2, 0)
 
 MATERIAL_POINTER_NAME = "vrml2_material"
-DATA_SCHEMA_VERSION = 2
+DATA_SCHEMA_VERSION = 3
 
 PREVIEW_NODE_TAG = "vrml2_preview_node"
 PREVIEW_ROLE_TAG = "vrml2_preview_role"
@@ -19,13 +19,11 @@ EXPORT_KEYS = {
     "initialized": "vrml2_initialized",
     "enabled": "vrml2_enabled",
     "def_name": "vrml2_defName",
-    "ambient_intensity": "vrml2_ambientIntensity",
     "diffuse_color": "vrml2_diffuseColor",
-    "include_emissive_color": "vrml2_includeEmissiveColor",
     "emissive_color": "vrml2_emissiveColor",
-    "shininess": "vrml2_shininess",
-    "include_specular_color": "vrml2_includeSpecularColor",
     "specular_color": "vrml2_specularColor",
+    "ambient_intensity": "vrml2_ambientIntensity",
+    "shininess": "vrml2_shininess",
     "transparency": "vrml2_transparency",
 }
 
@@ -43,15 +41,15 @@ META_KEYS = {
 }
 
 VRML_DEFAULTS = {
-    "ambient_intensity": 0.2,
     "diffuse_color": (0.8, 0.8, 0.8),
     "emissive_color": (0.0, 0.0, 0.0),
-    "shininess": 0.2,
     "specular_color": (0.0, 0.0, 0.0),
+    "ambient_intensity": 0.2,
+    "shininess": 0.2,
     "transparency": 0.0,
 }
 
-VRML_FIELD_INCLUSION_DEFAULTS = {
-    "include_emissive_color": False,
-    "include_specular_color": False,
-}
+LEGACY_EXPORT_KEYS = (
+    "vrml2_includeEmissiveColor",
+    "vrml2_includeSpecularColor",
+)